[tracing] Add support for preventing users from changing table sort order
This patch adds a userCanModifySortOrder property to <tr-ui-b-table>. By
default, it is set to true. If set to false, the user will NOT be able
to change the table's sort order by clicking on buttons in the column
headers.

This is useful for tables displaying ordered information (e.g. the
current stack trace). Note that it is still possible to change the sort
order by setting sortColumnIndex and sortDescending properties of the
table programmatically.

BUG=catapult:#2751

Review-Url: https://codereview.chromium.org/2328963002