Display zero values in value-set-table
So, it turns out that it is possible for telemetry to generate an empty
results2.html. Instead of throwing an exception deep in tr-ui-b-table when
value-set-table tries to select the first cell in the 1-th column, which doesn't
exist, let's move the "zero values" message from metrics-side-panel to
value-set-table so that it also shows up in results2.html.

runMetrics() is split out of metricMapFunction so that metrics-side-panel doesn't
need to suppress IterationInfo and deserialize the histograms that are serialized by metricMapFunction, which was breaking related-event-set-span.
value-set-table's groupby-picker's possibleGroups is now re-computed from
DEFAULT_POSSIBLE_GROUPS every time the values are changed (actually every time
updateContents_() is called, but that's a separate bug).
maybeDisableShowAll_ doesn't need to inspect every value, it can stop if it finds
one that isn't in sourceValues.
show_all is no longer automatically unchecked every time updateContents_ is
called (which happens when the user checks show_all).

BUG=catapult:#2795

Review-Url: https://codereview.chromium.org/2347693002