[eslint] Add a custom rule for Catapult-style camelcase
Basically, this is the same as the eslint camelcase rule with a few
exceptions:

  - The suffixes _smallerIsBetter and _biggerIsBetter are allowed (for
    units)
  - The prefix g_ is allowed (for globals)
  - The prefix opt_ is allowed (for optional parameters)

I added a new folder for these rules at common/eslint, with me and Ethan
as the OWNERS of the folder (for now). Given that these were the same
OWNERS for the .eslintrc, it seemed like it made sense to move that
config into this folder, which I also did.

BUG=catapult:#2809
R=benjhayden@chromium.org, eakuefner@chromium.org, nednguyen@google.com

Review URL: https://codereview.chromium.org/2351173003 .