Display Histograms in value-set-table-cells.
Currently, value-set-table displays a single histogram below the table.
This histogram is below the fold when there are many rows in the table.
Histograms, their statistics, diagnostics, and sample diagnostics should all be
in the value-set-table's cells.

This patch adds a dom-module tr-v-ui-value-set-table-cell that manages the
transitions between scalar-span and histogram-span.

Screenshot (google-only): https://screenshot.googleplex.com/6W2a0wVjAxi.png

Value-set-table is still the only consumer of histogram-span, so this patch takes
some liberties with it:
Histogram-span now uses a BarChart instead of a ColumnChart so that bins are
comparable between columns of the value-set-span.
Histogram-level diagnostics are also now displayed in histogram-span.

Review-Url: https://codereview.chromium.org/2341623002