Entry point for bisect comparison.
This is meant to enable bisect to call stats logic in catapult. The
entrypoint will be tracing/bin/compare_samples, it will take paths to
the output of the test runs for each sample and a metric string as given
by the dashboard. This patch is meant to support chartjson, valueset
and buildbot result line output formats.

BUG=chromium:616932
R=eakuefner,benjhayden

Review-Url: https://codereview.chromium.org/2089833002