[code health] Create common/ and services/ folders within dashboard.
Services are modules for interacting with external web services. They represent
external dependencies, so they generally have few dependencies within dashboard.
The addition of the bisect module causes it makes sense to have another folder
within dashboard for this kind of common code.

Common contains all code that will be common to both modules. dashboard/ can
depend on common/, but not the other way around.

The end-state dependency graph is supposed to look like this:
- common/ has no dependencies on any other folders.
- services/ can only depend on common/.
- models/ can depend on common/ and services/.
- dashboard/ can depend on all of the above.

There are a bunch of violations of these rules in unit tests :(
But none in the non-test code, I think :)

BUG=catapult:#2891

Review-Url: https://codereview.chromium.org/2384973002