[tr-ui-b-table] Prevent line breaks around expand-buttons.
Currently, when a tr-ui-b-table is wider than the window, the browser notices
that the expand-buttons have display:inline-block, so it saves horizontal space
by breaking the line after them.

Before: https://screenshot.googleplex.com/44YHsSPXddQ.png
After: https://x20web.corp.google.com/~benjhayden/Screenshot%202016-10-04%20at%209.39.29%20PM.png

This CL sets td.style.display = 'flex' on cells that contain an <expand-button>
in order to prevent this line break.

This made the margins look funny, so this CL also fixes BASIC_INDENTATION to
respond to whatever the font-size of the table is.

tr-ui-b-table cannot set display:flex in CSS for all <td> elements because that would break table layout. It seems to be ok to set it for one <td> in each <tr>.
There is currently no way to select the <td> elements that contain an
<expand-button> in CSS since there is no parent selector.

BUG=catapult:#2902

Review-Url: https://codereview.chromium.org/2398553002