Wait a frame before finishing navigation
** PERF SHERIFF: this CL may cause a small regression to
telemetry benchmark but it's expected ***

When running a telemetry benchmark, we don't want to start
interaction gestures until the page is fully interactive. Currently,
the NavigationAction waits until document.readyState is at least
interactive. This isn't enough, since it only indicates that Blink
is ready; the page's layers may not yet have made it through the
compositor's commit cycle yet. We need to wait until the compositor
is ready since the gestures may well be handled on the compositor
thread.

BUG=chromium:652905

Review-Url: https://codereview.chromium.org/2400123002