Fix histogram-span to use NameBarChart.
Currently, histogram-span uses BarChart, and sets the height of each bar
proportional to the range of the bin that it represents.
The problem is that bin ranges can vary by orders of magnitude, producing
a few very large bars and many bars that are so small that they are invisible.

This CL makes all of the bars the same height regardless of their range
using NameBarChart to effectively set custom tick labels, though NameBarChart
thinks of them as bar names, since it doesn't use ticks.

This required sanitizing NameBarChart, which currently is-a ColumnChart.
NameBarChart is renamed NameColumnChart, and a new NameBarChart is added as a
subclass of BarChart.

When the height of the NameBarChart is not enough to contain all of the bar
labels, then some of them are elided instead of letting them overlap.
If callers do not want bar names to be elided, then they should set the height of
the chart to margin.top+margin.bottom+((1+parseInt(fontSize)+1)*bars.length).

Also, histogram-span now sets the color of its bars to "blue" instead of the
hash of the string "y", which was an ugly brown, since literally every histogram
on image search is blue, in the millenial sense of the word "literally", natch.

https://screenshot.googleplex.com/TmYW5kKvCu0.png

BUG=catapult:#2859

Review-Url: https://codereview.chromium.org/2399013004