Speed up value-set-table...rebuild.
console.time() showed that updateGroups_ was not a bottleneck, but
rebuilding the tr-ui-b-table alone was taking over 1s on the new 'performance'
test case, which pretends to run 4 pages with story-repeat=5, each run of which
produces 100 histograms.

This CL speeds up rebuilding the table by 90% to 100ms by making
value-set-table-cell delay creating the histogram-span until the user wants to
display it.

Review-Url: https://codereview.chromium.org/2403603002