[eslint] Don't enforce eslint's "eqeqeq" rule for dashboard/ code
Dashboard, at least for now, uses == and != the correct way: as a means
to check for loose equality. "Fixing" all instances of this would be
error-prone and time-consuming, and we don't have consensus that it's
the right thing to do.

See https://codereview.chromium.org/2402513002/#msg7 for more discussion
on this.

Doing this required moving back to using directory-based .eslintrc
configs instead of the --config .eslintrc flag.

I thought it would be nice to be able to keep all of the .eslintrc
configuration in a single directory, but in reality it takes away our
ability to have per-directory .eslintrc overrides to the global
configuration because any rules specified through --config supercede
directory-specific rules (see:
http://eslint.org/docs/user-guide/configuring).

This CL reduces our existing eslint violations for 202 to 24.

TBR=sullivan@chromium.org
BUG=catapult:#2915

Review-Url: https://codereview.chromium.org/2409153002