Improve usability of groupby-picker.
There are 2 kinds of discoverability issues with the current groupby-picker:
You need to click Add another in order to see disabled possible groupings, and
You need to hover over a grouping and pay close attention to see its close button
and figure out that the hand cursor suggests draggability.

This CL changes the groupby-picker away from drag-drop to use more discoverable
checkboxes and buttons.

https://screenshot.googleplex.com/6SabSQRCAYE.png

All available grouping keys are always visible.
No hover effects.
No mysterious cursor changes.
This reduces the number of clicks required to enable a disabled grouping key from 2 to 1.
This simplifies the gestures required to reorder grouping keys from a
complicated mousedown-drag-mouseup gesture to 1 click, or click-move-click, or
click-move-click-move-click, etc.

Buttons that have no effect (like the left/right buttons for disabled groupings)
are hidden in order to limit visual noise, but there are still lots of buttons
when there are lots of enabled categories, which can happen in histogram-set-table.
We'll keep a look out for emerging web designs that are more minimalist while
maintaining the discoverability of buttons and checkboxes.

Review-Url: https://codereview.chromium.org/2458873002