Allow toggling name column max-width in histogram-set-table.
Currently, the name column is hard-coded to 300px maximum.
This is too small for memory metric names.
I haven't yet figured out a good way to use a drag-handle, or else clip the
beginning of the name ("reverse ellipsis").
This CL adds a button to the name header cell to allow disabling the max-width
to show the full name.

http://www/~benjhayden/2946.png

BUG=catapult:#2946

Review-Url: https://codereview.chromium.org/2484113002