Dashboard - Bisect bot guessing code should return valid bot.
Previously if the guessing code couldn't match the bot name to a bisect bot, it
defaulted to 'linux_perf_bisect' which causes problems on other masters when
that bot doesn't exist. Bisects initiated would immediately fail with an invalid
config error.

Instead, we check if linux_perf_bisect is available, and fall back to using any
bot from the available bots list.

Maybe in the future we should have the concept of a specific fallback bot per master?

BUG=catapult:#664484

Review-Url: https://codereview.chromium.org/2534653002