Enable arrow-body-style eslint rule
This rule requires arrow functions that return literals to use the
literal syntax. The exception to this is object literals, which require
a return statement to clarify the ambiguous () => {} problem, where it's
not clear whether we're returning an empty object literal or just
creating a no-op function. (ES6 treats this as a no-op function.)

TBR=sullivan@chromium.org
BUG=catapult:#3039

Review-Url: https://codereview.chromium.org/2533833002