[Telemetry] Increase timeout after sending SIGTERM
To shut down the browser on desktop, we first try to quit the browser normally,
then send SIGTERM, then wait a while, then send SIGKILL if we still don't see
that the browser is shut down.

Apparently, with CHROME_HEADLESS=1, we should be getting a .dmp file on
SIGTERM, but we see no such file.

kbr@ suggests that this may simply be because we're not waiting long enough.
The best result of this CL would be that we start getting useful stack traces
from these launch failures.

BUG=catapult:#3074

Review-Url: https://codereview.chromium.org/2568033003