Dashboard - Support build failures in bisect output.
Previously the recipe filtered the json data and the dashboard just outputted whatever was there. With this CL the dashboard will selectively print out build failures from the revision data depending on whether they're in the culprit range.

Note: Not much of the output should change at first, only thing that should change is when we push the json data changes to prod you should see this for build failures:

===== TESTED REVISIONS =====
Revision  Mean  Std Dev  N    Good?
100       100   0        10   good
102       100   0        10   good
103       ---   ---      ---  build failure
104       ---   ---      ---  build failure
105       200   0        10   bad
106       200   0        10   bad

Or in the case of lots and lots of failures:
===== TESTED REVISIONS =====
Revision  Mean  Std Dev  N    Good?
100       100   0        10   good
102       100   0        10   good
103       ---   ---      ---  build failure
---       ---   ---      ---  too many build failures to list
114       ---   ---      ---  build failure
115       200   0        10   bad
116       200   0        10   bad

BUG=chromium:668540, chromium:665439

Review-Url: https://codereview.chromium.org/2588143002