Unit refactor cl 2.  Add UnitScale
This is part of a larger Unit/UnitScale/TimeDisplayModes refactor.
It adds UnitScales, which include the entire unit symbol instead of just
the prefix symbol. This allows us to add a scale for time that includes
units larger than seconds. This will in turn allow us to add a Unit that can format
to times larger than seconds.

BUG=catapult:#3098

Review-Url: https://codereview.chromium.org/2615993003