Make extract_thread_list slightly smarter
If adb is killed or restarted in response
to issuing the ps dump then extract_thread_list
can crash trying to int(cols[1]) as it assumes
that the first line is the PS header.

Instead, have it search for the first line that
contains USER at the beginning and treat that
as the header line. This avoids treating any
messages from adb as part of the ps dump.

Fixes b/34127555

Review-Url: https://codereview.chromium.org/2638333003