Add Segments to the UserModel.
Currently, the UserModel contains only UserExpectations.
Due to the messy nature of reality, UserExpectations can overlap.
This can make them difficult for some metrics to process.

This CL defines an abstraction layer atop UserExpectations: Segments cannot
overlap, and provide helper methods for metrics to decide how to handle a
Segment.
Metrics can ask whether a Segment intersects any of a set of UE types and/or
initiatorTypes, or metrics can ask whether a Segment intersects *only* any of a set
of UE types and/or initiatorTypes.

If multiple UEs overlap, then the Segment that represents the time range during
which the UEs overlap will contain references to all of the overlapping UEs.

BUG=catapult:#3155

Review-Url: https://codereview.chromium.org/2585483002