Redesign breakdown-span.
Currently, breakdown-span contains a ColumnChart with a legend for Histograms
and well-behaved numbers, plus a drag-handle to resize the ColumnChart, plus a
table for weird numbers.
It often defaults to the wrong size.
It uses old ColorScheme colors, except for metrics that used
ChromeUserFriendlyCategory, which get the SinebowColorGenerator treatment.
Chart legend keys are not hoverable.

Tufte suggests that simple tables of numbers are good enough more often than we
think.

This new design for breakdown-span replaces the ColumnChart's legend with a
smarter table that can contain Histograms, well-behaved numbers, as well as weird
numbers.
The table contains percentages as well as absolute numbers.
The table is sorted sensibly, with Histograms and well-behaved numbers preceding
weird numbers, breaking ties by the name of the entry.
Instead of supporting resizing, the chart's height is fixed at 130px so that all
charts in a histogram-set-table are the same size and can be visually compared.
The table's height is fixed at 150px with overflow-y:auto to accomodate v8's
huge breakdowns.
Instead of allowing hovering legend keys to display one datum at a time, all of
the data is always displayed in the table so that it can be compared between
Histograms.
Hovering over rects in the BarChart highlights corresponding rows in the table
instead of displaying the chart's default hover box.
SinebowColorGenerator is used directly for metrics that did not specify a
colorScheme.

http://i.imgur.com/8o7IbDX.png

BUG=catapult:#3045,catapult:#3044,catapult:#3043,catapult:#2852

Review-Url: https://codereview.chromium.org/2560953002