Make the whole story power metric not depend on Chrome trace.
Currently, the "whole story" power metric (story:power) only looks at
data within th eboundaries of the Chrome trace, and returns no data
if there is nothing identifying the trace as being for Chrome. This CL
chagnes it to have this metric look at the entire trace, regardless
of the presence or absence of Chrome data.

BUG=catapult:#3106

Review-Url: https://codereview.chromium.org/2644733006