Store CPU usage data on the model via an auditor.
The purpose of this change is to separate the code that generates the CPU usage data from the code that displays it. This change will enable CPU usage data collected this way to be used in metrics directly rather than just displayed.

This change also paves the way for introducing additional data series similar to the CPU usage data series, such as GPU usage series.

Review-Url: https://codereview.chromium.org/2590553002