Fix related histogram links.
Currently, related-histogram-set-span, related-histogram-map-span, and
breakdown-span for RelatedHistogramBreakdowns display analysis-links whose
selection points to individual Histograms. When the analysis-links are clicked,
they fire a requestSelectionChange event. When histogram-set-table catches one of
these events, it tries to show that Histogram, enabling Show All if necessary.

However, there is not currently any way for users to narrow the displayed
HistogramSet to only all Histograms in a RelatedHistogramX.

This CL adds 'Show All' analysis-links to those spans mentioned above, and changes
the analysis-links' selections to arrays of Histogram names. When
histogram-set-table catches a requestSelectionChange event, it sets its search
query to a regexp that will match only Histograms whose names are in the
analysis-link's selection. It also focuses the search box to make it easy for users
to clear or modify the query.

This should make it much easier to manage which histograms are displayed.

http://i.imgur.com/6K0gScS.png

http://i.imgur.com/MpAqLkc.png

This is probably not the optimal UX. Maybe eventually we'll find time for user
studies, but we need something now.

Review-Url: https://codereview.chromium.org/2671453002