[tracing] Cache view builders in heap dump pane
In the heap dump details view switching repeatedly between view modes
is slow because we unnecessarily rebuild our model of the stack frame/type tree
every time we switch. By cacheing the model we can avoid rebuilding it
which can save seconds when switching back and forth on large traces.

BUG=catapult:#3162

Review-Url: https://codereview.chromium.org/2669913003