Refactor TraceData to encapsulate internal traces' representation
Changes to TraceData's API:
+ TraceData.Serialize() is the only supported way to serialize the traces to a single
HTML trace file. The memory saving still comes from the fact that this uses the trace
file handle directly if the trace were collected using TraceFileHandle (in
chrome_inspector/tracing_backend.py).
+ TraceData.GetTracesFor only return traces in dictionary form, since one do not
needs to call this unless they want to have access to the trace events in Python.
Upon calling this, all the TraceFileHandle are also converted to the in-memory form
automatically since there is no use in saving memory once this is invoked.

As a result of this, no clients of TraceData need concern with whether the internal
representation of trace is TraceFileHandle or the in-memory dictionary form.

BUG=catapult:#3119

Review-Url: https://codereview.chromium.org/2661573003