Add TEAM and COMPONENT to catapult OWNERS files.
This follows the practics outlined here:
https://docs.google.com/document/d/1jty6UsFMW9-SYgpQC-ztEc3lltziOQBBArMkROCSTh0/edit#heading=h.xgjl2srtytjt

In addition to having a consistent place for people to find contact
information, chromium's tooling will also be updated to use these files
for setting bug components. For example, FindIt may automatically add
Speed>Telemetry to a bug it files when it believes the culprit CL is in
the catapult/telemetry directory.

This CL is dependent on benhenry@'s changes to the speed bug components:
https://docs.google.com/document/d/19dpu0Si0Tg7JOvVuPOoFyrdfwSmrmsV4IHNhdQ11IMo/edit#heading=h.kdwz1bz4knz6

I wanted to send it out for early comments in case this flushes out
more discussion points around the bug components change.

Review-Url: https://codereview.chromium.org/2665323004