CollectedGeneric is a better way to merge Generic Diagnostics.
Currently, Generic uses a hacky way to merge.
This CL breaks merging Generics but prepares to use a better way that will also
work for other Diagnostics: cloneEmpty() will come in a follow-up CL.
This CL is split out of https://codereview.chromium.org/2683463002

The name "Collected" indicates that the Generics are not being "merged" in the
same sense that Histograms are merged, and are rather being "collected" atomically,
as a preview of #2854.

BUG=catapult:#2842,catapult:#2854

Review-Url: https://codereview.chromium.org/2697773002