Deduplicate Diagnostics.
This is an optimization both for [de]serialization and for merging Histogram
Relationships: https://codereview.chromium.org/2683463002
When merging Histogram Relationships, leaf histograms may be merged to multiple
other Histograms. In order to figure out which of those should be related to the
Histogram that owns the current RelatedHistogram Diagnostic, the TelemetryInfos,
DeviceInfos, and BuildbotInfos must be compared. That could be done using their
equals() methods, but that would impact results.html responsiveness far too much.
Merging Relationships will be significantly faster if it can compare object
identity and rely on deduplication.

BUG=catapult:#2842

Review-Url: https://codereview.chromium.org/2692863003