Adds UE histograms as a diagnostics to the total histogram in runtimeStatsTotalMetric.
runtimeStatsTotalMetric computes runtimeStats metrics and bucket them based
on the user expectations (UEs). It also adds the histograms in the "Any"
bucket that contains data for all the samples. This cl uses adds the
histograms on the UE buckets as a diagnostic to the corresponding histograms
in the Any bucket. This helps organize the data better and also hides the
histograms from the UE buckets in the main results page.

BUG=chromium:686250

Review-Url: https://codereview.chromium.org/2698583005