Update the update_v8 script and roll d8 on Windows to stable
Changes in the last year to the way that v8 handles date
internationalization have made it so that we need to pull in the ICU
data file that helps v8 understand how to stringify dates in different
locales.

I verified that this didn't break any tests by running:

  $ python tracing/bin/run_vinn_tests

and verifying that all tests passed. (They didn't before fixing the
internationalization problems.)

NOTRY=true
BUG=catapult:#3189
R=nednguyen@google.com

Review-Url: https://codereview.chromium.org/2708153002 .