[Telemetry] Fix subtle bug in android_dumpsys_power_monitor
The previous version used the result of platform.RunCommand (which
returns a string) as the input to csv.reader (which expects an iterable
of lines). The csv.reader was thus iterating over each character
in the output as if it were a line.

Using device.RunShellCommand (which returns a list of lines) fixes
the issue.

BUG=catapult:#3308

Review-Url: https://codereview.chromium.org/2721553005