Do not auto-open histograms in singular rows.
Currently, if there is only a single row, histogram-set-table automatically
opens its Histograms.
This is a UX optimization: we expect users to decide which cells to open and
then click them, so if we know which cells they will want to open (because there
are so few of them to choose from), then we don't have to wait for them to
pick and click, we can just go ahead and automatically open them.

In practice, most metrics produce more than one Histogram, so the single-row
situation only arises in blink_perf.dom tests.
Recent feedback has called into question the value of automatically overloading
the user with the full contents of every Histogram, so this CL removes this UX
optimization.

Review-Url: https://codereview.chromium.org/2745413003