Add tr-v-ui-histogram-set-controls.
This new element will replace the top controls that are currently in
histogram-set-table.
In the next CL, these controls will be deleted from histogram-set-table, and
histogram-set-view will contain both a h-s-controls and a h-s-table.

This element is a View/Controller for the top half of the
HistogramSetConfiguration MVC Model.

This CL depends on https://codereview.chromium.org/2745983004
This CL is split out from https://codereview.chromium.org/2747453003

BUG=catapult:#3289

Review-Url: https://codereview.chromium.org/2747083003