[Systrace][BattOr] Simplify BattOr options in systrace.
Before it had the complete set of options to do automapping from hub1 to hub2
like the labs use, but these options drastically complicate interacting with
systrace and the BattOr with no benefit gained. This does not fix the already
broken BattOr functionality in systrace. That will be done in this cl:
https://codereview.chromium.org/2712163002/

BUG=catapult:#3262

Review-Url: https://codereview.chromium.org/2736123003