Always upload BattOr serial logs to cloud storage
We need this as a debugging tool to debug http://crbug.com/699581
because we need both the preceding and subsequent serial log for each
failure serial log.

One case of _UploadSerialLogToCloudStorage was left untouched: when
_SendBattOrCommand fails. The reason I left this untouched was that if
that command fails, we already upload the serial log to cloud storage.
If it succeeds, then we'll eventually reach CollectTraceData, where I
changed it to unconditionally upload to cloud storage.

BUG=chromium:699581

Review-Url: https://codereview.chromium.org/2750963003