[devil] Provide device.GetApplicationPids()
A convenient method for clients to get PIDs when they know the exact
application (package) name of the process they want to get, without
having to inspect and validate the return value of GetPids().

Also provides an |at_most_one| option to raise an exception if many
PIDs are found but at most one was expected.

BUG=catapult:#3378

Review-Url: https://codereview.chromium.org/2744373003