Add WALT trace agent
This CL creates a new trace agent for the WALT device
(github.com/google/walt). This trace agent imports events that come from
both the physical WALT device and the WALT Android app. These events are
timestamped with the CLOCK_MONOTONIC times. This CL also extends
ftrace_importer.html to allow clocks other than the ftrace_global clock.
The clock id is read from a comment in the trace of a specific format.

Review-Url: https://codereview.chromium.org/2774353003