[Telemetry] Enable more useful error stack for native crash with faulthandler tool
Motivation: the bug in https://github.com/catapult-project/catapult/issues/3525
was very difficult to figure out because there was no error stack. It's possible
that it was due to a native crash. Using faulthandler tool, we have a much
better stack in https://pastebin.com/5mZEDpuc that could have helped us debug
the issue much easier.

Review-Url: https://codereview.chromium.org/2848743002